Adding resource for homebrew installation
I've been asked a few times from folks at Turing across TPS about how to solve the Homebrew installation error that occurs when using a Mackbook running on a Apple Silicon (M1 or M2 chipset). 

I wrote a short quickstart guide as it may be able to be shared with others.